LPS-38484 Do not show "display-date" column in view "View History" for default value Web Contents
=== Current behavior
Since LPS-30512, we don't save the displayDate for a Web Content that stores the default values of a structure. Therefore, the Display Date column will always be empty in the view View History when your are editing this default content.

=== Expected behavior
Display this column only for "real" Web Contents.


== About the changes
It is safe to remove an element from list "headerNames" and from map "orderableHeaders", as they are instance-scoped variables of a SearchContainer object:

private List<String> _headerNames;
private Map<String, String> _orderableHeaders;

The static initializations in the ArticleSearch class ensure that the newly created container instances will have the same initial names, but when we call #add, #clear or #remove the in views, these actions change the instance variables, not the static ones.

Removing "display-date" elements in certain case (viewing history of a default article content) seems more reasonable than clearing the whole list/map and refill it with the same values this one.